misc updates to get this working again
The last update was in 2013, and the last travis test pass was in 2013 too.
The karma config file was working for grunt-karma 0.4, but was failing to find phantomjs.
Updating the karma config to a more recent module.exports... format along with grunt-karma version 2 fixed that
Otherwise there was also a small bug fix in a andCallThrough function call in the test